fix: change type int to double for divide

+++ Program.cs
@@ -7,13 +7,13 @@
             Console.WriteLine("Hesap Makinesi");
 
             Console.WriteLine("Birinci Sayıyı Giriniz: ");
-            int sayi1 = int.Parse(Console.ReadLine());
+            double sayi1 = double.Parse(Console.ReadLine());
 
             Console.WriteLine("İşlem Seçiniz (+) (-) (*) (/) (%): ");
             string islem = Console.ReadLine();
 
             Console.WriteLine("İkinci Sayıyı Giriniz: ");
-            int sayi = int.Parse(Console.ReadLine());
+            double sayi = double.Parse(Console.ReadLine());
 
 
             switch (islem)
